requirements.txt
python-docx
PyPDF2
python-dotenv
redis
Flask
requests